+++ Model/Client/DTO/Payment/CreatePaymentWebhook.php
@@ -10,10 +10,12 @@ class CreatePaymentWebhook extends AbstractRequest
     const EVENT_PAYMENT_RESERVATION_CREATED = 'payment.reservation.created';
     const EVENT_PAYMENT_CHECKOUT_COMPLETED = 'payment.checkout.completed';
     const EVENT_PAYMENT_CHARGE_CREATED = 'payment.charge.created';
+    const EVENT_PAYMENT_NEW_CHARGE_CREATED = 'payment.charge.created.v2';
     const EVENT_PAYMENT_CHARGE_FAILED = 'payment.charge.failed';
     const EVENT_PAYMENT_REFUND_INITIATED = 'payment.refund.initiated';
+    const EVENT_PAYMENT_NEW_REFUND_INITIATED = 'payment.refund.initiated.v2';
     const EVENT_PAYMENT_REFUND_FAILED = 'payment.refund.failed';
-    const EVENT_PAYMENT_REFUND_COMPLETED= 'payment.refund.completed';
+    const EVENT_PAYMENT_REFUND_COMPLETED = 'payment.refund.completed';
     const EVENT_PAYMENT_CANCEL_CREATED = 'payment.cancel.created';
     const EVENT_PAYMENT_CANCEL_FAILED = 'payment.cancel.failed';
 
@@ -24,11 +26,13 @@ class CreatePaymentWebhook extends AbstractRequest
         self::EVENT_PAYMENT_CREATED => 'SaveOrder',
         self::EVENT_PAYMENT_RESERVATION_CREATED => 'ReservationCreated',
         self::EVENT_PAYMENT_CHECKOUT_COMPLETED => 'CheckoutCompleted',
-        self::EVENT_PAYMENT_CHARGE_CREATED => '',
+        self::EVENT_PAYMENT_CHARGE_CREATED => 'PaymentCharged',
+        self::EVENT_PAYMENT_NEW_CHARGE_CREATED => 'PaymentCharged',
         self::EVENT_PAYMENT_CHARGE_FAILED => '',
-        self::EVENT_PAYMENT_REFUND_INITIATED => '',
+        self::EVENT_PAYMENT_REFUND_INITIATED => 'PaymentRefund',
+        self::EVENT_PAYMENT_NEW_REFUND_INITIATED => 'PaymentRefund',
         self::EVENT_PAYMENT_REFUND_FAILED => '',
-        self::EVENT_PAYMENT_REFUND_COMPLETED => '',
+        self::EVENT_PAYMENT_REFUND_COMPLETED => 'PaymentRefund',
         self::EVENT_PAYMENT_CANCEL_CREATED => '',
         self::EVENT_PAYMENT_CANCEL_FAILED => '',
     ];